+++ libtelemetry/src/test/java/com/mapbox/android/telemetry/TelemetryClientLocationEventTest.java
@@ -26,7 +26,7 @@ public class TelemetryClientLocationEventTest extends MockWebServerTest {
     when(mockedContext.getSystemService(Context.ACTIVITY_SERVICE)).thenReturn(mockedActivityManager);
 
     TelemetryClient telemetryClient = obtainATelemetryClient("anyAccessToken", "anyUserAgent",
-      mockedContext);
+      "reformedUserAgent", mockedContext);
     double aLatitude = 40.416775;
     double aLongitude = -3.703790;
     Event aLocationEvent = new LocationEvent("aSessionId", aLatitude, aLongitude, "");